NerdShit
Database Config like halfway done ish. Implemented some of the addChallenge and showChallenge methods, as well as removeChallenge. All that remains is the rest of the add- and show- challenge methods as well as adding and removing different objects separately, but that shouldn't take much time to implement, since the rest of the things look fairly similar to the already implemented ones.